docker/python/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
requests==2.25.1
Werkzeug==2.0.3
gunicorn==20.0.4
mysql-connector-python==8.0.28
SQLAlchemy==1.4.41
